add book-details
