+++ src/app/blog-details/blog-details.component.html
@@ -1,5 +1,4 @@
 <div>
-
   <!-- Start Bradcaump area -->
   <div class="ht__bradcaump__area bg-image--6">
     <div class="container">
