+++ src/app/blog/blog.component.html
@@ -28,7 +28,7 @@
               <h2>Category Archives: HTML</h2>
             </div>
             <!-- Start Single Post -->
-            <article class="blog__post d-flex flex-wrap">
+            <article class="blog__post d-flex flex-wrap" >
               <div class="thumb">
                 <a href="blog-details.html">
                   <img src="assets/images/blog/blog-3/1.jpg" alt="blog images">
