--- src/app/board-admin/board-admin.component.html
@@ -1,5 +1,0 @@
-<div class="container">
-  <header class="jumbotron">
-    <p>{{ content }}</p>
-  </header>
-</div>
