--- src/app/board-admin/board-admin.component.ts
@@ -1,24 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { UserService } from '../_services/user.service';
-
-@Component({
-  selector: 'app-board-admin',
-  templateUrl: './board-admin.component.html',
-  styleUrls: ['./board-admin.component.css']
-})
-export class BoardAdminComponent implements OnInit {
-  content = '';
-
-  constructor(private userService: UserService) { }
-
-  ngOnInit() {
-    this.userService.getAdminBoard().subscribe(
-      data => {
-        this.content = data.message;
-      },
-      err => {
-        this.content = JSON.parse(err.error).message;
-      }
-    );
-  }
-}
