--- src/app/board-moderator/board-moderator.component.ts
@@ -1,24 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { UserService } from '../_services/user.service';
-
-@Component({
-  selector: 'app-board-moderator',
-  templateUrl: './board-moderator.component.html',
-  styleUrls: ['./board-moderator.component.css']
-})
-export class BoardModeratorComponent implements OnInit {
-  content: any;
-
-  constructor(private userService: UserService) { }
-
-  ngOnInit() {
-    this.userService.getModeratorBoard().subscribe(
-      data => {
-        this.content = data.message;
-      },
-      err => {
-        this.content = JSON.parse(err.error).message;
-      }
-    );
-  }
-}
