--- src/app/board-user/board-user.component.html
@@ -1,6 +1,0 @@
-<div class="container">
-  <header class="jumbotron">
-    <a href="/create" routerLink="create">Create the book</a>
-    <p>{{ content }}</p>
-  </header>
-</div>
