--- src/app/board-user/board-user.component.ts
@@ -1,25 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { UserService } from '../_services/user.service';
-
-@Component({
-  selector: 'app-board-user',
-  templateUrl: './board-user.component.html',
-  styleUrls: ['./board-user.component.css']
-})
-export class BoardUserComponent implements OnInit {
-  content = '';
-
-  constructor(private userService: UserService) { }
-
-  ngOnInit() {
-    this.userService.getUserBoard().subscribe(
-      data => {
-        console.log(data);
-        this.content = data.message;
-      },
-      err => {
-        this.content = JSON.parse(err.error).message;
-      }
-    );
-  }
-}
