--- src/app/components/book-details/book-details.component.css
@@ -1,2 +1,0 @@
-
-
