--- src/app/components/book-details/book-details.component.html
@@ -1,20 +1,0 @@
-<h2>Book Details</h2>
-
-<hr/>
-<div *ngIf="book">
-  <div>
-    <label><b>Name: </b></label> {{book.name}}
-  </div>
-  <div>
-    <label><b>Description: </b></label> {{book.description}}
-  </div>
-  <div>
-    <label><b>Price: </b></label> {{book.unitPrice}}
-  </div>
-</div>
-
-<br>
-<br>
-<button (click)="list()" class="btn btn-primary">Back to Book List</button><br>
-
-
