--- src/app/components/book-list/book-list.component.html
@@ -1,27 +1,0 @@
-<div class="row">
-
-    <!--<div class="col-lg-4 col-md-6 mb-4" *ngFor="let book of books | async">
-      <div class="card h-100">
-        <a href="#">
-          <img class="card-img-top img-responsive" src="{{book.imageUrl}}" height="250" width="100">
-        </a>
-        <div class="card-body">
-          <h4 class="card-title">
-            <a href="#">{{book.name}}</a>
-          </h4>
-          <h5>{{book.unitPrice | currency: 'INR'}}</h5>
-          <p class="card-text text-truncate">{{book.description}}</p>
-        </div>
-        <div class="card-footer">
-          <div class="col-lg-4 col-md-6 mb-4">
-            <button (click)="deleteBook(book.id)" class="btn btn-danger" *ngIf="showDeleteButton">Delete</button>
-            <button (click)="updateBook(book.id)" class="btn btn-info" *ngIf="showUpdateButton">
-              Update
-            </button>
-            <button (click)="bookDetails(book.id)" class="btn btn-info">Details</button>
-          </div>
-        </div>
-      </div>
-    </div>-->
-
-</div>
