--- src/app/components/create-book/create-book.component.html
@@ -1,25 +1,0 @@
-<h3>Create Book</h3>
-<div [hidden]="submitted" style="width: 400px;">
-  <form (ngSubmit)="onSubmit()">
-    <div class="form-group">
-      <label for="name">Name</label>
-      <input type="text" class="form-control" id="firstName" required [(ngModel)]="book.name" name="name">
-    </div>
-
-    <div class="form-group">
-      <label for="name">Description</label>
-      <input type="text" class="form-control" id="lastName" required [(ngModel)]="book.description" name="description">
-    </div>
-
-    <div class="form-group">
-      <label for="name">Price</label>
-      <input type="text" class="form-control" id="emailId" required [(ngModel)]="book.unitPrice" name="Price">
-    </div>
-    <button type="submit" class="btn btn-success">Submit</button>
-  </form>
-</div>
-
-<div [hidden]="!submitted">
-  <h4>You submitted successfully!</h4>
-  <!-- <button class="btn btn-success" (click)="newBook()">Add</button> -->
-</div>
