--- src/app/components/create-book/create-book.component.ts
@@ -1,42 +1,0 @@
-import { Book } from '../../common/book';
-import { BookService } from '../../_services/book.service';
-import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
-
-@Component({
-  selector: 'app-create-Book',
-  templateUrl: './create-book.component.html',
-  styleUrls: ['./create-book.component.css']
-})
-export class CreateBookComponent implements OnInit {
-
-  book: Book = new Book();
-  submitted = false;
-
-  constructor(private bookService: BookService,
-              private router: Router) { }
-
-  ngOnInit() {
-  }
-
-  newBook(): void {
-    this.submitted = false;
-    this.book = new Book();
-  }
-
-  save() {
-    this.bookService.createBook(this.book)
-      .subscribe(data => console.log(data), error => console.log(error));
-    this.book = new Book();
-    this.gotoList();
-  }
-
-  onSubmit() {
-    this.submitted = true;
-    this.save();
-  }
-
-  gotoList() {
-    this.router.navigate(['/books']);
-  }
-}
