--- src/app/components/page-not-found/page-not-found.component.css
@@ -1,6 +1,0 @@
-body {
-    background: #dedede;
-}
-.page-wrap {
-    min-height: 75vh;
-}
