+++ src/app/home/home.component.html
@@ -63,8 +63,8 @@
           <div class="product product__style--3">
             <div class="col-lg-3 col-md-4 col-sm-6 col-12">
               <div class="product__thumb">
-                <a class="first__img" href="single-product.html"><img src="assets/images/books/1.jpg" alt="product image"></a>
-                <a class="second__img animation1" href="single-product.html"><img src="assets/images/books/2.jpg" alt="product image"></a>
+                <a class="first__img" routerLink="" (click)="bookDetails(book.id)"><img src="assets/images/books/1.jpg" alt="product image"></a>
+                <a class="second__img animation1" routerLink="" (click)="bookDetails(book.id)"><img src="assets/images/books/2.jpg" alt="product image"></a>
                 <div class="hot__box">
                   <span class="hot-label">BEST SALLER</span>
                 </div>
