+++ src/app/shop/shop.component.ts
@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import {Observable} from "rxjs";
+import {Book} from "../common/book";
+import {BookService} from "../_services/book.service";
 
 @Component({
   selector: 'app-shop',
@@ -6,9 +9,12 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./shop.component.css']
 })
 export class ShopComponent implements OnInit {
+  books: Observable<Book[]>;
 
-  constructor() { }
+  constructor(private bookService: BookService) { }
 
-  ngOnInit(): void {}
+  ngOnInit(): void {
+    this.books = this.bookService.getBookList();
+  }
 
 }
